pyproject.toml
[project]
name = "vector-search-using-faiss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "numpy>=2.2.6",
    "sentence-transformers>=5.1.0",
]
